requirements.txt
Pillow==9.4.0
Pillow==10.0.0
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.20.0
